requirements.txt
flask==2.2.5
requests==2.31.0
beautifulsoup4==4.14.3
psycopg2-binary==2.9.9
cloudinary==1.44.1
python-dotenv==0.21.1
gunicorn==21.2.0
